+++ src/Vortice.WinUI/Composition/ICompositorInterop.cs
@@ -7,19 +7,15 @@ using Microsoft.UI.Composition;
 namespace Vortice.WinUI.Composition;
 
 [Guid("FAB19398-6D19-4D8A-B752-8F096C396069")]
-public partial class ICompositorInterop : ComObject
+public partial class ICompositorInterop(nint nativePtr) : ComObject(nativePtr)
 {
-    public ICompositorInterop(IntPtr nativePtr) : base(nativePtr)
-    {
-    }
-
-    public static explicit operator ICompositorInterop?(IntPtr nativePtr) => nativePtr == IntPtr.Zero ? null : new ICompositorInterop(nativePtr);
+    public static explicit operator ICompositorInterop?(nint nativePtr) => nativePtr == 0 ? null : new ICompositorInterop(nativePtr);
 
     public unsafe Result CreateGraphicsDevice(IUnknown device, out CompositionGraphicsDevice? resultDevice)
     {
-        IntPtr devicePtr = MarshallingHelpers.ToCallbackPtr<IUnknown>(device);
-        IntPtr outDevice = default;
-        Result result = ((delegate* unmanaged[Stdcall]<IntPtr, void*, IntPtr*, int>)this[3])(NativePointer, (void*)devicePtr, &outDevice);
+        nint devicePtr = MarshallingHelpers.ToCallbackPtr<IUnknown>(device);
+        nint outDevice = default;
+        Result result = ((delegate* unmanaged[Stdcall]<nint, void*, nint*, int>)this[3])(NativePointer, (void*)devicePtr, &outDevice);
         if (result.Failure)
         {
             resultDevice = default;