+++ src/Vortice.WinUI/ISwapChainPanelNative.cs
@@ -2,30 +2,26 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repository root for more information.
 
 using Vortice.DXGI;
-#if WINDOWS
-using WinRT;
-using Microsoft.UI.Xaml.Controls;
-#endif
 
 namespace Vortice.WinUI;
 
 [Guid("63aad0b8-7c24-40ff-85a8-640d944cc325")]
 public unsafe partial class ISwapChainPanelNative : ComObject
 {
-    public ISwapChainPanelNative(IntPtr nativePtr)
+    public ISwapChainPanelNative(nint nativePtr)
         : base(nativePtr)
     {
     }
 
-    public static explicit operator ISwapChainPanelNative?(IntPtr nativePtr) => nativePtr == IntPtr.Zero ? null : new ISwapChainPanelNative(nativePtr);
+    public static explicit operator ISwapChainPanelNative?(nint nativePtr) => nativePtr == 0 ? default : new ISwapChainPanelNative(nativePtr);
 
 #if WINDOWS
-    public ISwapChainPanelNative(SwapChainPanel owner)
-        : base(((IWinRTObject)owner).NativeObject.GetRef())
+    public ISwapChainPanelNative(Microsoft.UI.Xaml.Controls.SwapChainPanel panel)
+        : base(WinUIHelpers.GetNativeObject(typeof(ISwapChainPanelNative).GUID, panel))
     {
     }
 
-    public static explicit operator ISwapChainPanelNative(SwapChainPanel owner) => new(owner);
+    public static explicit operator ISwapChainPanelNative(Microsoft.UI.Xaml.Controls.SwapChainPanel panel) => new(panel);
 #endif
 
     public IDXGISwapChain SwapChain